Description
This is a simple game where the user is asked to guess a number from a random range. Here, the lower range is always 1 but the upper range can be anywhere from 3 to 20, which is randomly generated. If the user guesses the number on his first try he will score a maximum of 10 points and a special message will generate. On the other hand, if he guesses higher or lower than the randomly generated number, 1 point is deducted from the possibility of 10 points. The guessing and point deduction continues until the user guesses the correct number.